refactor: clean up user profile pieces
