+++ src/routers/auth/profile.ts
@@ -1,21 +1,19 @@
 import prisma from "@mirlo/prisma";
 import { NextFunction, Request, Response } from "express";
 
-import {
-  USER_PROFILE_SELECT,
-  serializeUserProfile,
-} from "../../serializers/userProfile";
+import { serializeUser } from "../../serializers/user";
+import { userSelect } from "../../utils/user";
 
 const profile = async (req: Request, res: Response, next: NextFunction) => {
   const { email } = req.user as { email: string };
   try {
     const foundUser = await prisma.user.findFirst({
       where: { email },
-      select: USER_PROFILE_SELECT,
+      select: userSelect,
     });
 
     res.status(200).json({
-      result: foundUser ? serializeUserProfile(foundUser) : null,
+      result: foundUser ? serializeUser(foundUser) : null,
     });
   } catch (e) {
     next(e);
